requirements.txt
discord
disputils
requests==2.24.0
aiohttp==3.6.2
pillow==7.2.0
randfacts
pymongo
dnspython
psutil
